вынес класс App в отдельный файл
